pyproject.toml
[project]
name = "job-hunter-agent"
version = "0.1.0"
description = "Autonomous multi-agent job discovery system"
requires-python = ">=3.12"
dependencies = [
    # Core
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # LLM
    "anthropic>=0.40",
    "instructor>=1.0",
    # Orchestration (for internal agent graphs)
    "langgraph>=0.2",
    # Scraping
    "crawl4ai>=0.4",
    "playwright>=1.40",
    # Search
    "tavily-python>=0.3",
    # PDF
    "docling>=2.0",
    "pdfplumber>=0.10",
    "pypdf>=4.0",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "pgvector>=0.3",
    # Caching
    "redis>=5.0",
    # Embeddings (local fallback)
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    # Output
    "openpyxl>=3.1",
    "pandas>=2.1",
    # Email
    "sendgrid>=6.0",
    "aiosmtplib>=3.0",
    "jinja2>=3.1",
    # CLI
    "typer>=0.9",
    "rich>=13.0",
    # Observability
    "structlog>=24.0",
    "langsmith>=0.1",
    "tenacity>=8.2",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=1.20",
    # Scheduling
    "apscheduler>=3.10",
    # HTTP client
    "httpx>=0.27",
    "email-validator>=2.0",
]

[project.scripts]
job-hunter = "job_hunter_cli.main:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "ASYNC", "B", "C4", "DTZ", "PT", "RUF"]
ignore = ["ANN101", "ANN102"]

[tool.ruff.lint.per-file-ignores]
"src/job_hunter_cli/main.py" = ["B008"]  # typer requires function calls in defaults

[tool.ruff.lint.isort]
known-first-party = ["job_hunter_core", "job_hunter_agents", "job_hunter_infra", "job_hunter_cli"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "tavily",
    "tavily.*",
    "sendgrid",
    "sendgrid.*",
    "crawl4ai",
    "crawl4ai.*",
    "pdfplumber",
    "pdfplumber.*",
    "pypdf",
    "pypdf.*",
    "sentence_transformers",
    "sentence_transformers.*",
    "docling",
    "docling.*",
    "aiosmtplib",
    "aiosmtplib.*",
    "apscheduler",
    "apscheduler.*",
    "langsmith",
    "langsmith.*",
    "pandas",
    "pandas.*",
    "openpyxl",
    "openpyxl.*",
    "instructor",
    "instructor.*",
    "opentelemetry",
    "opentelemetry.*",
]
ignore_missing_imports = true
follow_imports = "skip"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short -v"
markers = [
    "unit: fast, fully mocked tests",
    "integration: requires real DB",
    "live: requires real network access",
    "slow: takes > 5 seconds",
]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_hunter_core", "src/job_hunter_agents", "src/job_hunter_infra", "src/job_hunter_cli"]
